% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{sdg_attributes}
\alias{sdg_attributes}
\title{Load available attributes from the UNSD SDG database}
\usage{
sdg_attributes(goal = NULL, series = NULL)
}
\arguments{
\item{goal}{Numeric or character value between 1 and 17 corresponding to
one of the
\href{https://www.un.org/development/desa/disabilities/envision2030.html}{17
SDGs}.}

\item{series}{Character value of one of the
\href{https://unstats.un.org/sdgs/indicators/database/}{SDG series codes}.}
}
\value{
A data frame.
}
\description{
\code{sdg_attributes()} provides a data frame of all available attributes
for a specific SDG goal or series of data.
}
