% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdg_data}
\alias{sdg_data}
\title{Load data from UNSD SDG database}
\usage{
sdg_data(series)
}
\arguments{
\item{series}{Character vector of
\href{https://unstats.un.org/sdgs/indicators/database/}{SDG series codes}.}
}
\value{
A data frame.
}
\description{
\code{sdg_data()} provides a data frame of all data for specified series
in the UNSD SDG database
}
