% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarea.R
\name{sdg_geoarea_data}
\alias{sdg_geoarea_data}
\title{Show SDG data available for specific geoArea}
\usage{
sdg_geoarea_data(area_code = NULL, area_name = NULL, returns = "all")
}
\arguments{
\item{area_code}{\href{https://unstats.un.org/unsd/methodology/m49/}{M49 code}
for a specific geoArea.}

\item{area_name}{\href{https://unstats.un.org/unsd/methodology/m49/}{M49 name}
for a specific geoArea. Recommended to use code instaed since the name
must exactly match.}

\item{returns}{String specifying the level of detail to return. "all" (default) and
"series" returns full details, while "indicators", "targets", and "goals"
provides details only to the specified level. Beyond that level, values are
summarized by their availability on the UN SDG database (e.g # of targets
available under each goal).}
}
\value{
A data frame.
}
\description{
\code{sdg_geoarea_data()} provides a data frame of all series
available for a specific \href{https://unstats.un.org/unsd/methodology/m49/}{
geoArea}. Must provide either an M49 area code or name, corresponding to
the UN SD standards. You can search for valid codes and names using
\code{sdg_geoareas()}.
}
