% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoarea.R
\name{sdg_geoareas}
\alias{sdg_geoareas}
\title{Load the geoArea reference table from the UNSD SDG database}
\usage{
sdg_geoareas(goal = NULL, target = NULL, indicator = NULL, series = NULL)
}
\arguments{
\item{goal}{Numeric or character value between 1 and 17 corresponding to
one of the
\href{https://www.un.org/development/desa/disabilities/envision2030.html}{17
SDGs}.}

\item{target}{Character value of one of the
\href{https://unstats.un.org/sdgs/indicators/Global\%20Indicator\%20Framework\%20after\%202020\%20review_Eng.pdf}{SDG
target codes} of the form "#.#".}

\item{indicator}{Character value of one of the
\href{https://unstats.un.org/sdgs/indicators/Global\%20Indicator\%20Framework\%20after\%202020\%20review_Eng.pdf}{SDG
indicator codes} of the form "#.#.#".}

\item{series}{Character value of one of the
\href{https://unstats.un.org/sdgs/indicators/database/}{SDG series codes}.}
}
\value{
A data frame.
}
\description{
\code{sdg_geoareas()} provides a data frame of all geoArea
\href{https://unstats.un.org/unsd/methodology/m49/}{M49 codes} and names.
If all arguments are NULL, returns all possible values, but if \code{goal},
\code{indicator}, or \code{series} specified, returns geoArea codes and names
available for that subset of the SDGs.
}
