% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{sdg_indicators}
\alias{sdg_indicators}
\title{Load available indicators from the UNSD SDG database}
\usage{
sdg_indicators(goals = NULL, targets = NULL)
}
\arguments{
\item{goals}{Numeric or character vector of values between 1 and 17
corresponding to the
\href{https://www.un.org/development/desa/disabilities/envision2030.html}{17
SDGs}.}

\item{targets}{Character vector of
\href{https://unstats.un.org/sdgs/indicators/Global\%20Indicator\%20Framework\%20after\%202020\%20review_Eng.pdf}{SDG
target codes} of the form "#.#".}
}
\value{
A data frame.
}
\description{
\code{sdg_indicators()} provides a data frame of all available
\href{https://unstats.un.org/sdgs/indicators/database/}{indicators in the
UNSD SDG database}.
}
