% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall.R
\name{sdg_overview}
\alias{sdg_overview}
\title{Load an overview of data available in the UNSD SDG database}
\usage{
sdg_overview(goals = 1:17, returns = "all")
}
\arguments{
\item{goals}{Numeric or character vector of values between 1 and 17
corresponding to the
\href{https://www.un.org/development/desa/disabilities/envision2030.html}{17
SDGs}.}

\item{returns}{String specifying the level of detail to return. "all" (default) and
"series" returns full details, while "indicators", "targets", and "goals"
provides details only to the specified level. Beyond that level, values are
summarized by their availability on the UN SDG database (e.g # of targets
available under each goal).}
}
\value{
A data frame.
}
\description{
\code{sdg_overview()} provides a data frame of all available
\href{https://unstats.un.org/sdgs/indicators/Global\%20Indicator\%20Framework\%20after\%202020\%20review_Eng.pdf}{goals,
targets, indicators, and series}, in the UNSD SDG database.
}
