% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{sdg_series}
\alias{sdg_series}
\title{Load all available series from the UNSD SDG database}
\usage{
sdg_series(all_releases = FALSE)
}
\arguments{
\item{all_releases}{If FALSE, the default, returns only the current release, and returns all releases if TRUE.}
}
\value{
A data frame.
}
\description{
\code{sdg_series()} provides a data frame of all available \href{https://unstats.un.org/sdgs/indicators/database/}{series in the UNSD SDG database}.
}
