% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{sdg_targets}
\alias{sdg_targets}
\title{Load available targets from the UNSD SDG database}
\usage{
sdg_targets(goals = NULL)
}
\arguments{
\item{goals}{Numeric or character vector of values between 1 and 17
corresponding to the
\href{https://www.un.org/development/desa/disabilities/envision2030.html}{17
SDGs}.}
}
\value{
A data frame.
}
\description{
\code{sdg_targets()} provides a data frame of all available
\href{https://unstats.un.org/sdgs/indicators/Global\%20Indicator\%20Framework\%20after\%202020\%20review_Eng.pdf}{targets
in the UNSD SDG database}.
}
