% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sdg_data}
\alias{sdg_data}
\title{Load data from UNSD SDG database}
\usage{
sdg_data(
  series,
  area_codes = NULL,
  time_start = NULL,
  time_end = NULL,
  clean_names = TRUE
)
}
\arguments{
\item{series}{Character vector of
\href{https://unstats.un.org/sdgs/indicators/database/}{SDG series codes}.}

\item{area_codes}{\href{https://unstats.un.org/unsd/methodology/m49/}{M49 codes}
for specific geoAreas.}

\item{time_start}{Numeric value indicating the starting (inclusive) year of
the series.}

\item{time_end}{Numeric value indicating the ending (inclusive) year of
the series.}

\item{clean_names}{Logical value indicating whether or not to clean up names
of the returned data frame. If \code{TRUE}, the default, spaces are replaced
with underscores and all other non-alphabetic characters are removed.}
}
\value{
A data frame.
}
\description{
\code{sdg_data()} provides a data frame of all data for specified series
in the UNSD SDG database.
}
